refactor pagination to separate component, fix sync
